requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.20.0
opencv-python>=4.5.0
pillow>=8.0.0
matplotlib>=3.5.0
scipy>=1.8.0
shapely>=1.8.0
tqdm>=4.60.0
numpy>=1.21.0 